move tests to "test" subpackage
